requirements.txt
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
AlexaPy==1.27.10
annotated-types==0.6.0
anyio==4.2.0
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
authcaptureproxy==1.3.2
azote==1.12.5
Babel==2.12.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.1.1
bleach==6.1.0
blinker==1.7.0
boto3==1.34.47
botocore==1.34.47
Bottleneck==1.3.7
Brlapi==0.8.5
Brotli==1.1.0
build==1.0.3
cairocffi==1.5.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dashdot==0.0.6
dbus-next==0.2.3
dbus-python==1.3.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
devscripts==2.23.5
distro==1.8.0
Django==5.0.2
dlib==19.24.2
dnf==4.19.2
docstring-to-markdown==0.15
docutils==0.20.1
envycontrol==3.4.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
fb-re2==1.0.7
file-magic==0.4.0
filelock==3.13.3
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.3.1
gensim==4.3.2
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.41
googlesearch-python==1.2.3
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
halo==0.0.31
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
humanize==3.13.1
hyperframe==6.0.1
i3ipc==2.2.1
idna==3.4
importlib-metadata==7.0.1
iniconfig==1.1.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.1.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
kaggle==1.6.8
keyring==24.3.0
kiwisolver==1.4.5
libcomps==0.1.20
libdnf==0.73.1
libvirt-python==9.7.0
lightgbm==4.3.0
log-symbols==0.0.14
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mercurial==6.5.3
meson==1.3.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nftables==0.1
nh3==0.2.15
notebook==7.1.0
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
olefile==0.46
opencv==4.8.1
openpyxl==3.1.0
overrides==7.7.0
packaging==23.1
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pefile==2023.2.7
pexpect==4.8.0
pillow==10.3.0
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.2.0
ply==3.11
progressbar2==3.53.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyalexa==0.1.1
pyarrow==15.0.0
pycairo==1.25.1
pycares==4.3.0
pycparser==2.20
pycups==2.0.1
pydantic==2.6.3
pydantic_core==2.16.3
pyelftools==0.31
pygame==2.5.2
pygit2==1.14.0
Pygments==2.17.2
PyGObject==3.46.0
pyotp==2.9.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.3.2
pytest-asyncio==0.21.0
python-augeas==1.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.1
python-slugify==8.0.4
python-utils==3.3.3
python-whois==0.9.3
python-xlib==0.33
pythonwhois==2.4.3
pytz==2024.1
pyudev==0.24.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readme-renderer==42.0
redis==5.0.3
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rpm==4.19.1.1
rpmautospec==0.6.3
rpmautospec-core==0.1.4
ruff==0.2.2
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
SCons==4.5.2
scylla-api-client==1.0
scylla-driver==3.26.4
seaborn==0.13.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
Send2Trash==1.8.2
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.33/src
setuptools==67.7.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
sos==4.6.0
soupsieve==2.5
spinners==0.0.24
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
staxlang==0.0.1
sympy==1.12
systemd-python==235
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traceback-with-variables==2.0.4
traitlets==5.9.0
treelib==1.7.0
twine==4.0.2
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.8.0
unidiff==0.7.5
uri-template==1.3.0
urllib3==1.26.18
urwid==2.1.2
uvicorn==0.27.1
virt-firmware==24.2
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xkbcommon==0.8
yarl==1.9.2
zipp==3.17.0
zombie-imp==0.0.2
